Added the volume dimension for 5D data
